infra/pyproject.toml
[project]
name = "tango-infra"
version = "0.1.0"
description = "AWS CDK infrastructure for Tango SaaS Starter"
requires-python = ">=3.12"
dependencies = [
    # Use latest CDK lib for Python 3.13 compatibility
    "aws-cdk-lib>=2.175.0",
    "constructs>=10.4",
    # Force latest jsii for Python 3.13 support
    "jsii>=1.106.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "boto3>=1.35.0",
    "psycopg2-binary>=2.9.9",
    "cfn-lint>=1.20.0",
    "ruff>=0.14",
]

[tool.ruff]
extend = "../ruff.toml"

[tool.ruff.lint.isort]
known-first-party = ["stacks"]
